workers/analytics-service/requirements.txt
aio_pika
psycopg2-binary